Refactor and decouple Payout logic
So that we can support lightning + external payout payments
