+++ BTCPayServer/Controllers/PullPaymentController.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
@@ -23,18 +24,21 @@ namespace BTCPayServer.Controllers
         private readonly CurrencyNameTable _currencyNameTable;
         private readonly PullPaymentHostedService _pullPaymentHostedService;
         private readonly BTCPayNetworkJsonSerializerSettings _serializerSettings;
+        private readonly IEnumerable<IPayoutHandler> _payoutHandlers;
 
         public PullPaymentController(ApplicationDbContextFactory dbContextFactory,
             BTCPayNetworkProvider networkProvider,
             CurrencyNameTable currencyNameTable,
             PullPaymentHostedService pullPaymentHostedService,
-            BTCPayServer.Services.BTCPayNetworkJsonSerializerSettings serializerSettings)
+            BTCPayServer.Services.BTCPayNetworkJsonSerializerSettings serializerSettings,
+            IEnumerable<IPayoutHandler> payoutHandlers)
         {
             _dbContextFactory = dbContextFactory;
             _networkProvider = networkProvider;
             _currencyNameTable = currencyNameTable;
             _pullPaymentHostedService = pullPaymentHostedService;
             _serializerSettings = serializerSettings;
+            _payoutHandlers = payoutHandlers;
         }
         [Route("pull-payments/{pullPaymentId}")]
         public async Task<IActionResult> ViewPullPayment(string pullPaymentId)
@@ -76,7 +80,7 @@ namespace BTCPayServer.Controllers
                               AmountFormatted = _currencyNameTable.FormatCurrency(entity.Blob.Amount, blob.Currency),
                               Currency = blob.Currency,
                               Status = entity.Entity.State.ToString(),
-                              Destination = entity.Blob.Destination.Address.ToString(),
+                              Destination = entity.Blob.Destination.ToString(),
                               Link = GetTransactionLink(_networkProvider.GetNetwork<BTCPayNetwork>(entity.Entity.GetPaymentMethodId().CryptoCode), entity.TransactionId),
                               TransactionId = entity.TransactionId
                           }).ToList()
@@ -95,11 +99,14 @@ namespace BTCPayServer.Controllers
             {
                 ModelState.AddModelError(nameof(pullPaymentId), "This pull payment does not exists");
             }
+            
             var ppBlob = pp.GetBlob();
             var network = _networkProvider.GetNetwork<BTCPayNetwork>(ppBlob.SupportedPaymentMethods.Single().CryptoCode);
-            IClaimDestination destination = null;
-            if (network != null &&
-                (!ClaimDestination.TryParse(vm.Destination, network, out destination) || destination is null))
+            
+            var paymentMethodId = ppBlob.SupportedPaymentMethods.Single();
+            var payoutHandler = _payoutHandlers.FirstOrDefault(handler => handler.CanHandle(paymentMethodId));
+            IClaimDestination destination = await payoutHandler?.ParseClaimDestination(paymentMethodId, vm.Destination);
+            if (destination is null)
             {
                 ModelState.AddModelError(nameof(vm.Destination), $"Invalid destination");
             }
