+++ BTCPayServer/Data/Payouts/BitcoinLike/IBitcoinLikeClaimDestination.cs
@@ -1,0 +1,9 @@
+using NBitcoin;
+
+namespace BTCPayServer.Data
+{
+    public interface IBitcoinLikeClaimDestination : IClaimDestination
+    {
+        BitcoinAddress Address { get; }
+    }
+}
