+++ BTCPayServer/Data/Payouts/IClaimDestination.cs
@@ -1,0 +1,7 @@
+namespace BTCPayServer.Data
+{
+    public interface IClaimDestination
+    {
+        
+    }
+}
