+++ BTCPayServer/EventAggregator.cs
@@ -111,6 +111,13 @@ namespace BTCPayServer
             s.Act = (o) => subscription(s, (T)o);
             return Subscribe(eventType, s);
         }
+        
+        public IEventAggregatorSubscription Subscribe(Type eventType,  Action<IEventAggregatorSubscription, object> subscription)
+        {
+            var s = new Subscription(this, eventType);
+            s.Act = (o) => subscription(s, o);
+            return Subscribe(eventType, s);
+        }
 
         private IEventAggregatorSubscription Subscribe(Type eventType, Subscription subscription)
         {
